Camera and mouseMoved fix

+++ GameFrame/src/NG/Camera/StaticCamera.java
@@ -65,7 +65,7 @@ public class StaticCamera implements Camera {
     }
 
     @Override
-    public void mouseMoved(int xDelta, int yDelta, int xPos, int yPos) {
+    public void mouseMoved(int xDelta, int yDelta, float xPos, float yPos) {
 
     }
