+++ GameFrame/src/NG/GUIMenu/Components/SDragEdge.java
@@ -47,7 +47,7 @@ public class SDragEdge extends SComponent implements MouseMoveListener {
     }
 
     @Override
-    public void mouseMoved(int xDelta, int yDelta, int xPos, int yPos) {
+    public void mouseMoved(int xDelta, int yDelta, float xPos, float yPos) {
         mouseMoved(xDelta, yDelta);
     }
 }
